Update user.c
I wrote the code but i dont know how to actually run it to check for errors.

Done in vscode so it dosent have syntax errors or forgotten semicolons   ;  at least .

The transfer part is the most incomplete one because i dont know how to verify the account number the user wants to transfer money to but will do later HOPEFULLY.

Review it once done in a very simple manner with a lot of comments so you can also understand. and let me know if you would like some changes.